Building Targetting and combat
Buildings can now be targetted and destroyed
Combat systems now also look for buildings

+++ Assets/Scripts/Damage Systems/DamageSystem.cs
@@ -3,5 +3,5 @@
 public abstract class DamageSystem : MonoBehaviour
 {
     public abstract void DealDamage();
-    public abstract void TakeDamage(float amount);
+    public abstract void TakeDamage(int amount);
 }
